bug: 파이어베이스 삭제 버그 원인 분석4

+++ src/main/java/com/be/inssagram/domain/post/service/PostService.java
@@ -143,13 +143,17 @@ public class PostService {
     public void deletePost(Long postId) {
         Post post = postRepository.findById(postId).orElseThrow(
                 PostDoesNotExistException::new);
-        List<String> fileNames = post.getFileNames();
-        for (String fileName : fileNames) {
-            fileName = String.format("%s/%s/%s",
-                    "post", post.getMember().getNickname(), fileName);
-            System.out.println(bucketName);
-            System.out.println(fileName);
-            firebaseStorageService.deleteFile(fileName);
+        if (post.getFileNames() != null) {
+            List<String> fileNames = post.getFileNames();
+            for (String fileName : fileNames) {
+                fileName = String.format("%s/%s/%s",
+                        "post", post.getMember().getNickname(), fileName);
+                System.out.println(bucketName);
+                System.out.println(fileName);
+                firebaseStorageService.deleteFile(fileName);
+            }
+        } else {
+
         }
         postRepository.save(post);  // 테스트 코드에서 확인 위한 작업.
         postRepository.delete(post);
